secure_file_sharing/requirements.txt
asgiref==3.8.1
cffi==1.17.1
coverage==7.6.10
cryptography==44.0.0
Django==4.2.17
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
itsdangerous==2.2.0
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyotp==2.9.0
pytest==8.3.4
pytest-cov==6.0.0
qrcode==8.0
sqlparse==0.5.3
Werkzeug==3.1.3
